refactor: extract the Avatar subcomponents into their own files
